+++ src/pages/Editor/index.tsx
@@ -1,11 +1,22 @@
+import * as monaco from 'monaco-editor';
+import { waitUntil } from '@/utils';
+import './worker';
 import './index.scss';
 
 const PageEditor: React.FC = () => {
-  console.log(1);
+  waitUntil(
+    () => document.getElementById('editor-container'),
+    (ele) => {
+      monaco.editor.create(ele, {
+        value: "function hello() {\n\talert('Hello world!');\n}",
+        language: 'javascript',
+      });
+    }
+  );
   return (
-    <>
-      <div className='page-editor'>Editor.</div>
-    </>
+    <div className='page-editor'>
+      <div id='editor-container' style={{ height: '100%' }}></div>
+    </div>
   );
 };