+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 # CHANGELOG
 
+## 23w4
+
+- add x6
+- add monaco editor